refactor(recordings): recording row actions barrelling with editing, transformation, and deletion capabilities
