+++ apps/app/src/routes/(config)/recordings/row-actions/index.ts
@@ -1,0 +1,1 @@
+export { default as RecordingRowActions } from './RecordingRowActions.svelte';
